+++ lib/card_input_flutter.dart
@@ -1,6 +1,5 @@
-library credit_card_input_flutter;
-
 export 'src/card_form.dart';
 export 'src/services/payment_service.dart';
 export 'src/services/result.dart';
 export 'src/models/card_model.dart';
+export 'src/models/card_form_config.dart';